🚚 move casting into atom.io
