+++ packages/anvl/tsconfig.json
@@ -1,9 +1,4 @@
 {
 	"extends": "../../tsconfig.json",
-	"include": [
-		"src",
-		"**/*.config.ts",
-		"../atom.io/__tests__/private/introspection/differ.test.ts",
-		"../atom.io/__tests__/private/introspection/refinery.test.ts"
-	]
+	"include": ["src", "**/*.config.ts"]
 }
